Version before adding saldo to entradasdetalle ans salidasdetalle
this version is previous to adding new field to tblEntradasDetalle
and tblSalidasDetalle to keep track of the balance

+++ SQLServerStudioProject/Larco/Larco/ALTER-tblEntradasDetalle.sql
@@ -19,10 +19,10 @@ ALTER TABLE tblEntradasDetalle
 ALTER COLUMN MAT_ID INT NOT NULL
 
 ALTER TABLE tblEntradasDetalle
-ALTER COLUMN ED_CANTIDAD INT NOT NULL
+ALTER COLUMN ED_CANTIDAD decimal(18,2) NOT NULL
 
 ALTER TABLE tblEntradasDetalle
-ALTER COLUMN ED_COSTO INT NOT NULL
+ALTER COLUMN ED_COSTO decimal(18,2) NOT NULL
 
 ---------------------------------------------
 --check this orders with rocio
